add new components, clean up slots screen implementation
