+++ src/App.svelte
@@ -2,15 +2,28 @@
   import { Router, Route } from "svelte-routing";
 
   import SlotsScreen from "~/screens/SlotsScreen.svelte";
+  import Slots from "~/screens/Slots/index.svelte";
+  import NavBar from "~/components/NavBar.svelte";
 
   export let url = window.location.pathname;
 </script>
 
-<div>
+<div class="container">
+  <NavBar />
   <Router {url}>
-    <Route path="/" component={SlotsScreen} />
+    <Route path="/" component={Slots} />
   </Router>
 </div>
 
 <style>
+  .container {
+    margin: auto;
+    width: 100%;
+    max-width: 800px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background: white;
+  }
 </style>
