--- src/components/TableRow.svelte
@@ -1,120 +1,0 @@
-<script>
-  export let gym, spaces, timing;
-  export let gymFilter, showAvailableOnly;
-
-  import { gymBookingLinks } from "~/constants";
-
-  const onRowClicked = () => {
-    window.open(gymBookingLinks[gym.slug], "_blank");
-  };
-</script>
-
-<tr
-  on:click={onRowClicked}
-  class:hidden={(gymFilter !== "all" && gymFilter !== gym.slug) ||
-    (showAvailableOnly && spaces < 1)}
-  class:warn={spaces < 10}
-  class:invalid={spaces < 1}
->
-  <td class="gym"
-    ><span
-      class:fitbloc={gym.slug === "fitbloc"}
-      class:boulderplus={gym.slug === "boulder-plus"}
-      class:bffclimb={gym.slug === "bff"}
-      class:oyeyo={gym.slug === "oyeyo"}
-      class:zvertigo={gym.slug === "z-vertigo"}
-      class:boulderworld={gym.slug === "boulder-world"}
-      class:lighthouse={gym.slug === "lighthouse"}
-      class:therockschool={gym.slug === "the-rock-school"}
-      class="badge">{gym.name}</span
-    ></td
-  >
-  <td class="timings">{timing}</td>
-  <td class="spaces"> <span class="badge">{spaces}</span></td>
-</tr>
-
-<style>
-  tr:hover,
-  tr:active {
-    background: aliceblue;
-  }
-
-  tr {
-    content-visibility: auto;
-    transition-duration: 0.3s;
-    cursor: pointer;
-  }
-
-  td {
-    font-size: 0.9em;
-    padding: 10px 5px;
-    border-top: solid 0.5px #f5f5f5;
-  }
-
-  td.spaces {
-    text-align: right;
-    padding-right: 10px;
-    min-width: 80px;
-  }
-
-  td.gym {
-    min-width: 120px;
-  }
-
-  .badge {
-    background: #f5f5f5;
-    padding: 4px 10px;
-    border-radius: 15px;
-    font-weight: bold;
-  }
-
-  .fitbloc {
-    background-color: #7b112a;
-    color: white;
-  }
-
-  .boulderplus {
-    background-color: #652580;
-    color: white;
-  }
-
-  .bffclimb {
-    background-color: #ea078d;
-    color: black;
-  }
-
-  .boulderworld {
-    background-color: #0c606b;
-    color: white;
-  }
-
-  .lighthouse {
-    background-color: #9e0102;
-    color: white;
-  }
-
-  .zvertigo {
-    background-color: yellow;
-    color: black;
-  }
-
-  .oyeyo {
-    background-color: #ef1e29;
-    color: white;
-  }
-
-  .therockschool {
-    background-color: black;
-    color: white;
-  }
-
-  .warn {
-    background-color: #e67d2243;
-  }
-
-  .invalid {
-    background: rgba(255, 175, 166, 0.284);
-    color: #e63946;
-    text-decoration: line-through;
-  }
-</style>
